cravecart-backend/requirements.txt
# ── Core ──────────────────────────────────────────────────
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
django-environ==0.11.2
Pillow==10.4.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# ── Auth ──────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1
django-allauth==64.2.1
social-auth-app-django==5.4.1

# ── AI ────────────────────────────────────────────────────
google-generativeai==0.7.2

# ── Async / Background Jobs ───────────────────────────────
celery==5.4.0
redis==5.0.8
django-redis==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# ── Email ─────────────────────────────────────────────────
django-anymail==11.1

# ── Storage ───────────────────────────────────────────────
django-storages==1.14.4
boto3==1.34.144

# ── Utilities ─────────────────────────────────────────────
ulid-py==1.1.0
django-extensions==3.2.3
django-model-utils==4.5.1
phonenumbers==8.13.43
python-dateutil==2.9.0.post0

# ── Production server ─────────────────────────────────────
gunicorn==22.0.0
whitenoise==6.7.0
sentry-sdk[django,celery,redis]==2.12.0
